Removing unused imports and files from the project.
