--- internal/render/testdata/dummychart/Chart.yaml
@@ -1,4 +1,0 @@
-apiVersion: v2
-name: dummy
-version: 0.1.0
-appVersion: "1.0.0"
