--- internal/render/testdata/dummychart/crds/dummy.example.com_dummyresources.yaml
@@ -1,20 +1,0 @@
-apiVersion: apiextensions.k8s.io/v1
-kind: CustomResourceDefinition
-metadata:
-  name: dummyresources.dummy.example.com
-spec:
-  group: dummy.example.com
-  names:
-    kind: DummyResource
-    plural: dummyresources
-  scope: Namespaced
-  versions:
-    - name: v1alpha1
-      served: true
-      storage: true
-      schema:
-        openAPIV3Schema:
-          type: object
-          properties:
-            spec:
-              type: object
