--- internal/render/testdata/dummychart/templates/_helpers.tpl
@@ -1,3 +1,0 @@
-{{- define "dummy.fullname" -}}
-{{- printf "%s-controller" .Release.Name -}}
-{{- end -}}
