--- internal/render/testdata/dummychart/templates/serviceaccount.yaml
@@ -1,7 +1,0 @@
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: {{ .Values.serviceAccount.name }}
-  namespace: {{ .Release.Namespace }}
-  annotations:
-    eks.amazonaws.com/role-arn: {{ index .Values.serviceAccount.annotations "eks.amazonaws.com/role-arn" }}
